Module Scfg

module Lexer : sig ... end

module Menhir_parser : sig ... end

module Parse : sig ... end

Module providing functions to parse a config from various kind of inputs.

module Pp : sig ... end

Module providing functions to pretty print a config or parts of a config.

module Query : sig ... end

Module providing functions to search inside a config.

module Types : sig ... end

Module defining types used to represent a config.