Module Scfg.Pp

Module providing functions to pretty print a config or parts of a config.

val param : Format.formatter -> string -> unit

Print a name or a parameter on a given formatter. The function will try to print with as low quoting as possible.

val params : Format.formatter -> string list -> unit

Print a list of parameters on a given formatter.

val children : Format.formatter -> Types.config -> unit

Print children of a directive on a given formatter.

val directive : Types.directive Fmt.t

Print a directive on a given formatter.

val config : Format.formatter -> Types.config -> unit

Print a config on a given formatter.